+++ CHANGELOG.md
@@ -27,6 +27,7 @@
 * Extract item view into a separate route. Fixes UIIN-955.
 * Filter instance, holding and item records by suppress: No. Refs UIIN-967, UIIN-968, UIIN-969, UIIN-970.
 * Filter instance by nature of content. Refs UIIN-824.
+* Filter instance by "Call number, eye readable" via holdings segment. Refs UIIN-858.
 
 ## [1.13.1](https://github.com/folio-org/ui-inventory/tree/v1.13.1) (2019-12-11)
 [Full Changelog](https://github.com/folio-org/ui-inventory/compare/v1.13.0...v1.13.1)